- Split FSM patches class into smaller classes - Added code to have Esdras move to a new position after defeat - Added code to move Perpetva apparition depending on Esdras position - Added code (WIP) to animate Esdras moving - Added code to turn both Esdras and Penitent around as needed
