+++ IterTormenti/IterTormenti.cs
@@ -66,9 +66,8 @@ namespace IterTormenti
         {
             if(newLevel == "D08Z01S01") // Bridge
             {
-                Esdras.FSMChanges.Apply();
+                Esdras.BossfightChanges.Apply();
             }
         }
-
     }
 }
